--- custom_cian/custom_cian/main_api/admin.py
@@ -1,3 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
