--- custom_cian/custom_cian/main_api/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class MainApiConfig(AppConfig):
-    name = 'main_api'
