+++ custom_cian/main/views.py
@@ -18,14 +18,16 @@ from django.http import Http404, \
                         HttpResponse, \
                         HttpResponseRedirect, \
                         HttpResponsePermanentRedirect
+from rest_framework import viewsets
+from rest_framework import permissions
 from django.shortcuts import render, redirect
-from django.template.response import TemplateResponse
 from django.views.decorators.csrf import csrf_protect
 from django.views.generic import DetailView, ListView
 from django.views.generic.edit import CreateView, UpdateView
 
 from .models import Realty, Tag, Saller, Subscriber, UserRealty
 from .forms import RealtyForm, SallerProfileForm
+from .serializers import UserSerializer, GroupSerializer
 
 from .tasks import send_email_task
 
@@ -220,3 +222,21 @@ class RealtyUpdateView(LoginRequiredMixin,
 
     def test_func(self) -> bool:
         return self.request.user.id == self.get_object().saller.created_by.id
+
+
+class UserViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows users to be viewed or edited.
+    """
+    queryset = User.objects.all().order_by('-date_joined')
+    serializer_class = UserSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+
+class GroupViewSet(viewsets.ModelViewSet):
+    """
+    API endpoint that allows groups to be viewed or edited.
+    """
+    queryset = Group.objects.all()
+    serializer_class = GroupSerializer
+    permission_classes = [permissions.IsAuthenticated]
